renamed FeuxDArtifice  to RdbmsIngestionComposition and added feux as alias
