+++ examples/pixdict/mysql_pixdict_ingestion_daily.py
@@ -1,7 +1,7 @@
 from airflow.configuration import conf
 
 from debussy_concert.data_ingestion.config.rdbms_data_ingestion import ConfigRdbmsDataIngestion
-from debussy_concert.data_ingestion.composition.feux_dartifice import FeuxDArtifice
+from debussy_concert.data_ingestion.composition.rdbms_ingestion import RdbmsIngestionComposition
 from debussy_concert.core.service.injection import inject_dependencies
 from debussy_concert.core.service.workflow.airflow import AirflowService
 
@@ -16,6 +16,6 @@ config_composition = ConfigRdbmsDataIngestion.load_from_file(
 
 inject_dependencies(workflow_service, config_composition)
 
-debussy_composition = FeuxDArtifice()
+debussy_composition = RdbmsIngestionComposition()
 
 dag = debussy_composition.auto_play()
